Add Note digital-twins-tutorial-sample-prereqs.md
Added Note for Visual Studio 2019 users who might be missing the installation workload needed for deploying an Azure Function.  I was able to validate the suggestion in this Note by first using the Visual Studio Code installer to uninstall the Azure Development Workload and then making an attempt to publish the SampleFunctionApp.